Merge in IN/integrationpoints from server-rc-REL-861575-rip-relone-cleanup to server-main

* commit '93155cb7daf9f8aa1a30c4c94452b91cd5523fb8': (57 commits)
  Disabled functional test run and one time setup log path
  Updated gold package Relativity.Sync
  Updated sutTemplate = "server2022-ga-rabbitmq" as suggested by Alphs team
  Updated latest Sync pre package
  REL-861575, latest IAPI gold package 2.9.1 is revved.
  Updated latest Relativity.Sync and Relativity.DataExchange.Client.SDK(IAPI)
  REL-861575: Re-enabled unit tests and no longer filter out functional tests depending on branch. The team will revisit functional test filters at a later point but, for now, want to ensure all changes made to RIP exerise all possible functional tests.
  REL-861575: Finalized the teardown/workspace delete issue with a TryDeleteWorkspace method.
  Added logs to identify Onetime tear down exception
  REL-861575: Removed RelOne K8s workload discovery service and associated tests.
  REL-861575: Added explicit IAPI/OI packages to ensure the assemblies + "oi" folder gets copied to the targetr directory.
  Commneted out  functional test filter selection logic. added nightly filters ti run.
  Added IIntegrationPointsAgentManager interface to avoid integrationpoints-kplerservice error.
  Revved gold packages of IntegrationPoints.MyFirstProvider, Relativity.IntegrationPoints.SDK and Relativity.IntegrationPoints.Services.Interfaces.Private
  Build fix
  - RelOne and Qgate functional tests related fixes - REG-CI removed these tests meant to be executed against our Regression-Environment, but the same tests are executed in CI/ApiTests - XssTests are not crucial at this moment - Ignored
  added relativityBranch property
  updated Relativity.sync and IntegrationPoints.MyFirstProvider packages
  Revved latest Relativity sync version.
  SERVER-RC Disable RelativityBranch in Tridentfile
  ...

+++ Source/kCura.IntegrationPoints.FunctionalTests/Functional/Helpers/RelativityFacadeExtensions.cs
@@ -41,11 +41,22 @@ namespace Relativity.IntegrationPoints.Tests.Functional.Helpers
             return instance.Resolve<IWorkspaceService>().Get(workspaceArtifactID);
         }
 
-        public static void DeleteWorkspace(this IRelativityFacade instance, Workspace workspace)
+        public static void TryDeleteWorkspace(this IRelativityFacade instance, Workspace workspace)
         {
             IWorkspaceService workspaceService = instance.Resolve<IWorkspaceService>();
+            try
+            {
+                if (!workspaceService.Exists(workspace.ArtifactID))
+                {
+                    return;
+                }
 
-            workspaceService.Delete(workspace.ArtifactID);
+                workspaceService.Delete(workspace.ArtifactID);
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine($"Functional test failed to delete workspace {workspace.ArtifactID}. Error: {e}");
+            }
         }
 
         public static void RequireAgent(this IRelativityFacade instance, string agentTypeName, int runInterval)
